Day4/pyproject.toml
[project]
name = "day4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.59.0",
    "google-generativeai>=0.8.5",
    "gradio>=5.38.2",
    "ipython>=9.4.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
]
